refs #12642 Refactored after rebase
