--- src/main/java/ru/nsu/ostest/adapter/in/rest/model/user/RoleDto.java
@@ -1,6 +1,0 @@
-package ru.nsu.ostest.adapter.in.rest.model.user;
-
-public record RoleDto(
-        String roleName
-) {
-}
