+++ src/main/java/ru/nsu/ostest/adapter/in/rest/model/user/role/RoleDto.java
@@ -1,4 +1,6 @@
 package ru.nsu.ostest.adapter.in.rest.model.user.role;
 
-public record RoleDto() {
+public record RoleDto(
+        String roleName
+) {
 }
