+++ src/main/java/ru/nsu/ostest/adapter/in/rest/model/user/userData/LogoutRequest.java
@@ -1,4 +1,4 @@
-package ru.nsu.ostest.adapter.in.rest.model.user;
+package ru.nsu.ostest.adapter.in.rest.model.user.userData;
 
 public record LogoutRequest(String accessToken, String refreshToken) {
 }
