+++ src/main/java/ru/nsu/ostest/security/AuthService.java
@@ -4,7 +4,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.NonNull;
 import ru.nsu.ostest.adapter.in.rest.model.user.auth.JwtResponse;
 import ru.nsu.ostest.adapter.in.rest.model.user.password.UserPasswordDto;
-import ru.nsu.ostest.adapter.in.rest.model.user.LogoutRequest;
+import ru.nsu.ostest.adapter.in.rest.model.user.userData.LogoutRequest;
 
 public interface AuthService {
