+++ src/test/java/ru/nsu/ostest/session/SessionControllerIntegrationTest.java
@@ -12,12 +12,9 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import ru.nsu.ostest.TransactionalHelper;
 import ru.nsu.ostest.adapter.in.rest.model.laboratory.LaboratoryCreationRequestDto;
-import ru.nsu.ostest.adapter.in.rest.model.session.GetLabSessionFroStudentRequestDto;
-import ru.nsu.ostest.adapter.in.rest.model.session.SessionDto;
-import ru.nsu.ostest.adapter.in.rest.model.session.StartSessionRequestDto;
+import ru.nsu.ostest.adapter.in.rest.model.session.*;
 import ru.nsu.ostest.adapter.in.rest.model.user.role.RoleEnum;
 import ru.nsu.ostest.adapter.in.rest.model.user.userData.UserCreationRequestDto;
-import ru.nsu.ostest.adapter.in.rest.model.session.*;
 import ru.nsu.ostest.adapter.mapper.LaboratoryMapper;
 import ru.nsu.ostest.adapter.mapper.UserMapper;
 import ru.nsu.ostest.adapter.out.persistence.entity.group.Group;
